[Challenges] - Bottom Navigation Bar
[Description]
It was made a big refactor so I can change the list value that's being shown to the challenge fragment.
There are no more two types of challenges. The only difference is the data that items will show from different API methods.
Still missing a way to change the list that's being observed.